+++ common/src/main/java/net/SpectrumFATM/black_archive/entity/client/ModModelLayers.java
@@ -20,6 +20,7 @@ public class ModModelLayers {
     public static final EntityModelLayer DALEK_SLAVE = new EntityModelLayer(new Identifier(BlackArchive.MOD_ID, "dalek_slave"), "main");
     public static final EntityModelLayer CYBERMAN = new EntityModelLayer(new Identifier(BlackArchive.MOD_ID, "cyberman"), "main");
     public static final EntityModelLayer CYBERMAT = new EntityModelLayer(new Identifier(BlackArchive.MOD_ID, "cybermat"), "main");
+    public static final EntityModelLayer ANGEL = new EntityModelLayer(new Identifier(BlackArchive.MOD_ID, "weeping_angel"), "main");
     public static final EntityModelLayer TIME_FISSURE = new EntityModelLayer(new Identifier(BlackArchive.MOD_ID, "time_fissure"), "main");
 
     public static final EntityModelLayer PILLAR_SHELL_LAYER = new EntityModelLayer(new Identifier(BlackArchive.MOD_ID, "pillar_shell"), "main");