+++ forge/src/main/java/net/SpectrumFATM/forge/entity/ModEntityAttributes.java
@@ -16,6 +16,7 @@ public class ModEntityAttributes {
         event.put(ModEntities.DALEK_PUPPET.get(), DalekPuppetEntity.createDalekSlaveAttributes().build());
         event.put(ModEntities.CYBERMAN.get(), CybermanEntity.createCyberAttributes().build());
         event.put(ModEntities.CYBERMAT.get(), CybermatEntity.createCyberAttributes().build());
+        event.put(ModEntities.ANGEL.get(), WeepingAngelEntity.createAngelAttributes().build());
         event.put(ModEntities.TIME_FISSURE.get(), TimeFissureEntity.createTimeFissureAttributes().build());
     }
 }