+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Forge Support!!!**
 - Tissue Compression Eliminator
 - Pillar Exterior
+- Weeping Angels
 
 ### Improved
 - Time rotor now continues to movie when in temporal orbit.